requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.2
openai==1.6.1
python-dotenv==1.0.1
requests==2.31.0
